chore(cms): move collections for better project organization
