+++ src/collections/BlogCategories/config.ts
