+++ src/collections/BlogPosts/config.ts
