+++ src/collections/BlogPosts/hooks/revalidatePost.ts
