+++ src/collections/Brands/config.ts
