+++ src/collections/FAQ/config.ts
