+++ src/collections/Locations/config.ts
