+++ src/collections/Pages/config.ts
