+++ src/collections/Work/config.ts
